+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-# 📦 CHANGELOG
+# CHANGELOG
 
-## [v0.1.0] - 2024-04-12
+## [v0.1.1] - 2024-04-12
 ### Added
 - Project structure and base modules
-- Basic user CRUD with SQLite
 - Initial README and documentation
+
+## [v0.1.2] - 2024-29-12
+- Added basic alembic config for DB migrations.
+- Initial models and schemas for User info